llm_evaluation_system/requirements.txt
nltk
rouge-score
openai
streamlit
python-dotenv
